change calc_product_rate change type of return value to integer

+++ RaspberryPi/Analysis/Analyse_Pulse_Signal/calc_product_rate.py
@@ -1,6 +1,6 @@
 # This program is written in Python3
 
 def calc_product_rate(count, interval):
-	product_rate = count / interval * 60
+	product_rate = int(count / interval * 60)
 
 	return calc_product
